Move installer factory, in preparation for upgrade detection.
